[nxp noup] platform: frdmrw612: Add support for frdmrw612
Add support for frdmrw612 in psa-arch-tests as per tfm v.2.1.
Copyright updates against NXP changes

The frdmrw612 can do attestation based on a HW key of a crypto
accelerator. For that situation it is not possible to know the key
beforehand, but for the verification of the initial attestation token
the public key needs to be exported from TF-M. This change is
implemented here.

Signed-off-by: Waqar Tahir <waqar.tahir@nxp.com>